Refactor, reformat, and rename into rerun_bridge
