--- requirements.txt
@@ -1,2 +1,0 @@
---pre -f https://build.rerun.io/commit/79d8021/wheels 
-rerun-sdk==0.13.0a3
