+++ rerun_bridge/package.xml
@@ -1,12 +1,13 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>rerun_logger</name>
+  <name>rerun_bridge</name>
   <version>0.1.0</version>
-  <description>The rerun_logger package</description>
+  <description>The rerun_bridge package</description>
   <maintainer email="opensource@rerun.io">rerun.io</maintainer>
   <license>Apache-2.0</license>
   <depend>roscpp</depend>
   <depend>sensor_msgs</depend>
+  <depend>nav_msgs</depend>
   <depend>cv_bridge</depend>
   <depend>geometry_msgs</depend>
   <depend>yaml-cpp</depend>
