+++ rerun_bridge/src/rerun_bridge/collection_adapters.hpp
@@ -1,0 +1,32 @@
+#pragma once
+
+#include <opencv2/opencv.hpp>
+#include <rerun.hpp>
+
+// Adapters so we can borrow an OpenCV image easily into Rerun images without copying:
+template <>
+struct rerun::CollectionAdapter<uint8_t, cv::Mat> {
+    /// Borrow for non-temporary.
+    Collection<uint8_t> operator()(const cv::Mat& img) {
+        assert(
+            "OpenCV matrix was expected have bit depth CV_U8" && CV_MAT_DEPTH(img.type()) == CV_8U
+        );
+
+        return Collection<uint8_t>::borrow(img.data, img.total() * img.channels());
+    }
+
+    // Do a full copy for temporaries (otherwise the data might be deleted when the temporary is destroyed).
+    Collection<uint8_t> operator()(cv::Mat&& img) {
+        assert(
+            "OpenCV matrix was expected have bit depth CV_U8" && CV_MAT_DEPTH(img.type()) == CV_8U
+        );
+
+        std::vector<uint8_t> img_vec(img.total() * img.channels());
+        img_vec.assign(img.data, img.data + img.total() * img.channels());
+        return Collection<uint8_t>::take_ownership(std::move(img_vec));
+    }
+};
+
+inline rerun::Collection<rerun::TensorDimension> tensor_shape(const cv::Mat& img) {
+    return {img.rows, img.cols, img.channels()};
+};
