--- rerun_logger/launch/example.launch
@@ -1,11 +1,0 @@
-<launch>
-  <!-- Publishes the voxel grid to rviz for display -->
-  <node pkg="rosbag" type="play" name="player" args="-s 11.0 -u 47.0 -r 2.0 $(find rerun_logger)/indoor_forward_3_snapdragon_with_gt.bag">
-  </node>
-
-  <!-- Run the logger node -->
-  <node name="rerun_logger_node" pkg="rerun_logger" type="rerun_logger_node" output="screen">
-    <rosparam param="yaml_path" subst_value="True">$(find rerun_logger)/launch/example_params.yaml</rosparam>
-  </node>
-
-</launch>
